% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.e.mcmc.R
\name{d.e.mcmc}
\alias{d.e.mcmc}
\title{Title}
\usage{
d.e.mcmc(f, max.iter, n.walkers, n.dim, init.range, ...)
}
\arguments{
\item{f}{a funtion ti sample frrom}

\item{max.iter}{}

\item{n.walkers}{}

\item{n.dim}{}

\item{init.range}{}

\item{...}{}
}
\value{
ir retur sfsdf
}
\description{
more detail akjdhakdh askjdjhsdfds
}
\examples{
1-1
}
