% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\docType{package}
\name{MCMCEnsembleSampler}
\alias{MCMCEnsembleSampler}
\alias{MCMCEnsembleSampler-package}
\title{MCMCEnsembleSampler}
\description{
This package implements a particle Monte Carlo Markov Chain sampler
with two different ways of creating proposals.
}
\author{
Sanda Kern-Dejanic
}
